= Moved `coordConvert` function into `utils`
